Another beautiful RGA output sample
This was was generated from 0 (random seeding). You can see the pure original improvisation by the model in the very beginning. And then it of course resolves into what it remembers once the spooky (latent) memory saturates. WOW!! Absolutely amazing if you ask me!!!

As always performance is incredible and it only takes ~10 minutes to generate 15 minutes (faster-than-real-time for music, lol)

Enjoy! :)